@@ -2,12 +2,12 @@ import { createContext, useState } from 'react';
 
 export const languageContext = createContext();
 
-export const LanguageProvider = ({children}) => {
+export const LanguageProvider = ({ children }) => {
   const [language, setLanguage] = useState('en');
 
-return(
-  <languageContext.Provider value={{language, setLanguage}}>
-    {children}
-  </languageContext.Provider>
-)
+  return (
+    <languageContext.Provider value={{ language, setLanguage }}>
+      {children}
+    </languageContext.Provider>
+  )
 }
