@@ -30,27 +30,27 @@ const ImgDiv = styled.div`
     }
   }
 `
-export default function ShowData({type,name,img1,img2,children}) {
+export default function ShowData({ type, name, img1, img2, children }) {
   const [hoverShow, setHoverShow] = useState(1);
   const [fullShow, setFullShow] = useState();
 
   useEffect(() => {
-    if(fullShow) {
-      document.documentElement.scrollTo({ top: document.documentElement.scrollTop+1, behavior: 'instant' });
+    if (fullShow) {
+      document.documentElement.scrollTo({ top: document.documentElement.scrollTop + 1, behavior: 'instant' });
       const preview = document.getElementById("fullShow");
-      preview.scrollIntoView({inline:"center",behavior:"smooth"});
+      preview.scrollIntoView({ inline: "center", behavior: "smooth" });
     }
-  },[fullShow])
-  
-  function bigPicture(e){
+  }, [fullShow])
+
+  function bigPicture(e) {
     setFullShow(e.target.src);
   };
-  
-  function exitPreview(){
+
+  function exitPreview() {
     setFullShow(false);
   }
 
-  return(
+  return (
     <div id="#home" css={css`
         display: grid;
         grid-template-columns: 1.5fr 1fr;
@@ -62,8 +62,10 @@ export default function ShowData({type,name,img1,img2,children}) {
           grid-template-columns: none;
         }
         `}>
-          {fullShow && <div id="fullShow" css={css`
+      {fullShow && <div id="fullShow" css={css`
             position: absolute;
+            /* position: fixed;
+            inset: 0px; */
             z-index: 2;
             display: flex;
             align-items: center;
@@ -72,23 +74,23 @@ export default function ShowData({type,name,img1,img2,children}) {
             height: 100vh;
             background-color: rgba(0,0,0,0.7);
             `}>
-              <div css={css`
+        <div css={css`
                 position: absolute;
                 right: 3%;
                 top: 3%;
                 cursor: pointer;
-                `} onClick={()=>exitPreview()}>
-                <img src={Cancel} alt='cancel'/>
-              </div>
-              <img css={css`height:85vh;
+                `} onClick={() => exitPreview()}>
+          <img src={Cancel} alt='cancel' />
+        </div>
+        <img css={css`height:85vh;
                 @media (max-width: 1300px) {
                   width: 100vw;
                   height: auto;
                 }
-                `} src={fullShow} alt='bigPicture'/>
-            </div>}
-        <ImgDiv>
-          <Img css={css`
+                `} src={fullShow} alt='bigPicture' />
+      </div>}
+      <ImgDiv>
+        <Img css={css`
             position: absolute;
             top: 17%;
             left: 3%;
@@ -98,18 +100,18 @@ export default function ShowData({type,name,img1,img2,children}) {
               left: 0;
             }
             `}
-            onClick={(e)=>bigPicture(e)}>
-            <img src={img1} alt='imagen1' width="500"/>
-          </Img>
-          <Img
-            onMouseOver={() => setHoverShow(-1)}
-            onMouseOut={() => setHoverShow(1)}
-            onClick={(e)=>bigPicture(e)}>
-            <img src={img2} alt='imagen2' width="500"/>
-          </Img>
-        </ImgDiv>
-        <div>
-          <div css={css`
+          onClick={(e) => bigPicture(e)}>
+          <img src={img1} alt='imagen1' width="500" />
+        </Img>
+        <Img
+          onMouseOver={() => setHoverShow(-1)}
+          onMouseOut={() => setHoverShow(1)}
+          onClick={(e) => bigPicture(e)}>
+          <img src={img2} alt='imagen2' width="500" />
+        </Img>
+      </ImgDiv>
+      <div>
+        <div css={css`
             font-size: 2.5rem;
             color:#5CB1B5;
             font-weight: 100;
@@ -118,24 +120,30 @@ export default function ShowData({type,name,img1,img2,children}) {
               font-size: 1.5rem;
             }
             `}>
-            {type}
-          </div>
-          <div css={css`
-            color: #5CB1B5;
+          {type}
+        </div>
+        <div css={css`
+            a:link,
+            a:visited,
+            a:active{
+              color: #5CB1B5;
+              text-decoration: none;
+            }
+            color: #5CB1B5!important;
             font-size: 3.5rem;
             @media (max-width: 768px){
               font-size: 2.3rem;
             }
             `}>{name}</div>
-          <div css={css`
+        <div css={css`
             font-size: 1.5rem;
             @media (max-width: 768px){
               font-size: 1.2rem;
             }
             `}>
-            {children}
-          </div>
+          {children}
         </div>
       </div>
+    </div>
   )
 }
